+++ changelog.md
@@ -12,6 +12,10 @@ Types of changes:
 - `Removed`    for now removed features.
 - `Fixed`      for any bug fixes.
 - `Security`    in case of vulnerabilities.
+
+Things to Check
+- setup.py version
+- readme?
 -->
 
 ## [Unreleased]